// src/payload-types.ts
export interface PayloadMedia {
  id: string
  url: string
  alt: string
  filename: string
}

export interface RichTextContent {
  root: {
    children: Array<{
      type?: string
      version?: number
      children: Array<{
        text: string
      }>
    }>
  }
}

export interface BaseItem {
  id: string
  createdAt: string
  updatedAt: string
}
// ...
export interface Temsilcilik extends BaseItem {
  firmaAdi: string
  ulke: string
  logo: PayloadMedia
  aciklama?: RichTextContent
  websitesi?: string
  siraNo?: number
}